Trying to fix failing submission-ui.spec.ts integration tests.
