--- cypress/integration/homepage-statistics.spec.ts
@@ -1,21 +1,0 @@
-import { testA11y } from 'cypress/support/utils';
-
-describe('Site Statistics Page', () => {
-    // NOTE: statistics were removed from the navbar
-    // it('should load if you click on "Statistics" from homepage', () => {
-    //     cy.visit('/');
-    //     cy.get('ds-navbar ds-link-menu-item a[title="Statistics"]').click();
-    //     cy.location('pathname').should('eq', '/statistics');
-    // });
-
-    it('should pass accessibility tests', () => {
-        cy.visit('/statistics');
-
-        // <ds-site-statistics-page> tag must be loaded
-        cy.get('ds-site-statistics-page').should('exist');
-
-        // TODO accessibility tests are failing because the UI has been changed
-        // Analyze <ds-site-statistics-page> for accessibility issues
-        // testA11y('ds-site-statistics-page');
-    });
-});
