+++ cypress/integration/submission.ts
@@ -1,0 +1,156 @@
+import { loginProcess } from '../support/commands';
+import { TEST_ADMIN_PASSWORD, TEST_ADMIN_USER, TEST_SUBMIT_COLLECTION_UUID } from '../support';
+
+const createItemProcess = {
+  // typeCollectionName() {
+  //   cy.get('.modal-body input[type = "search"]').type(collectionName);
+  // },
+  // selectCollection() {
+  //   cy.get('.modal-body .scrollable-menu button[title = "' + collectionName + '"]').eq(0).click();
+  // },
+  checkLocalHasCMDIVisibility() {
+    cy.get('#traditionalpageone form div[role = "group"] label[for = "local_hasCMDI"]').should('be.visible');
+  },
+  checkIsInputVisible(inputName, formatted = false, inputOrder = 0) {
+    let inputNameTag = 'input[';
+    inputNameTag += formatted ? 'ng-reflect-name' : 'name';
+    inputNameTag += ' = ';
+
+    cy.get('#traditionalpageone form div[role = "group"] ' + inputNameTag + '"' + inputName + '"]')
+      .eq(inputOrder).should('be.visible');
+  },
+  checkIsNotInputVisible(inputName, formatted = false, inputOrder = 0) {
+    let inputNameTag = 'input[';
+    inputNameTag += formatted ? 'ng-reflect-name' : 'name';
+    inputNameTag += ' = ';
+
+    cy.get('#traditionalpageone form div[role = "group"] ' + inputNameTag + '"' + inputName + '"]')
+      .eq(inputOrder).should('not.be.visible');
+  },
+  clickOnSelectionInput(inputName, inputOrder = 0) {
+    cy.get('#traditionalpageone form div[role = "group"] input[name = "' + inputName + '"]').eq(inputOrder).click();
+  },
+  clickOnInput(inputName, force = false) {
+    cy.get('#traditionalpageone form div[role = "group"] input[ng-reflect-name = "' + inputName + '"]')
+      .click(force ? {force: true} : {});
+  },
+  writeValueToInput(inputName, value, formatted = false, inputOrder = 0) {
+    if (formatted) {
+      cy.get('#traditionalpageone form div[role = "group"] input[ng-reflect-name = "' + inputName + '"]').eq(inputOrder).click({force: true}).type(value);
+    } else {
+      cy.get('#traditionalpageone form div[role = "group"] input[name = "' + inputName + '"]').eq(inputOrder).click({force: true}).type(value);
+    }
+  },
+  blurInput(inputName, formatted) {
+    if (formatted) {
+      cy.get('#traditionalpageone form div[role = "group"] input[ng-reflect-name = "' + inputName + '"]').blur();
+    } else {
+      cy.get('#traditionalpageone form div[role = "group"] input[name = "' + inputName + '"]').blur();
+    }
+  },
+  clickOnTypeSelection(selectionName) {
+    cy.get('#traditionalpageone form div[role = "group"] div[role = "listbox"]' +
+      ' button[title = "' + selectionName + '"]').click();
+  },
+  clickOnSuggestionSelection(selectionNumber) {
+    cy.get('#traditionalpageone form div[role = "group"] ngb-typeahead-window[role = "listbox"]' +
+      ' button[type = "button"]').eq(selectionNumber).click();
+  },
+
+  clickOnDivById(id, force) {
+    cy.get('div[id = "' + id + '"]').click(force ? {force: true} : {});
+  },
+  checkInputValue(inputName, observedInputValue) {
+    cy.get('#traditionalpageone form div[role = "group"] div[role = "combobox"] input[name = "' + inputName + '"]')
+      .should('contain',observedInputValue);
+  },
+  checkCheckbox(inputName) {
+    cy.get('#traditionalpageone form div[role = "group"] div[id = "' + inputName + '"] input[type = "checkbox"]')
+      .check({force: true});
+  },
+  controlCheckedCheckbox(inputName, checked) {
+    const checkedCondition = checked === true ? 'be.checked' : 'not.be.checked';
+    cy.get('#traditionalpageone form div[role = "group"] div[id = "' + inputName + '"] input[type = "checkbox"]')
+      .should(checkedCondition);
+  },
+  clickOnSave() {
+    cy.get('.submission-form-footer button[id = "save"]').click();
+  },
+  clickOnSelection(nameOfSelection, optionNumber) {
+    cy.get('.dropdown-menu button[title="' + nameOfSelection + '"]').eq(optionNumber).click();
+  },
+  clickAddMore(inputFieldOrder) {
+    cy.get('#traditionalpageone form div[role = "group"] button[title = "Add more"]').eq(inputFieldOrder)
+      .click({force: true});
+  },
+  checkDistributionLicenseStep() {
+    cy.get('ds-clarin-license-distribution').should('be.visible');
+  },
+  checkDistributionLicenseToggle() {
+    cy.get('ds-clarin-license-distribution ng-toggle').should('be.visible');
+  },
+  checkDistributionLicenseStatus(statusTitle: string) {
+    cy.get('div[id = "license-header"] button i[title = "' + statusTitle + '"]').should('be.visible');
+  },
+  clickOnDistributionLicenseToggle() {
+    cy.get('ds-clarin-license-distribution ng-toggle').click();
+  },
+  checkLicenseResourceStep() {
+    cy.get('ds-submission-section-clarin-license').should('be.visible');
+  },
+  clickOnLicenseSelectorButton() {
+    cy.get('ds-submission-section-clarin-license div[id = "aspect_submission_StepTransformer_item_"] button').click();
+  },
+  checkLicenseSelectorModal() {
+    cy.get('section[class = "license-selector is-active"]').should('be.visible');
+  },
+  pickUpLicenseFromLicenseSelector() {
+    cy.get('section[class = "license-selector is-active"] ul li').eq(0).dblclick();
+  },
+  checkLicenseSelectionValue(value: string) {
+    cy.get('ds-submission-section-clarin-license select[id = "aspect_submission_StepTransformer_field_license"]').contains(value);
+  },
+  selectValueFromLicenseSelection(option: string) {
+    cy.get('ds-submission-section-clarin-license select[id = "aspect_submission_StepTransformer_field_license"]').select(option);
+  },
+  checkResourceLicenseStatus(statusTitle: string) {
+    cy.get('div[id = "clarin-license-header"] button i[title = "' + statusTitle + '"]').should('be.visible');
+  },
+  showErrorMustChooseLicense() {
+    cy.get('div[id = "sectionGenericError_clarin-license"] ds-alert').contains('You must choose one of the resource licenses.');
+  },
+  showErrorNotSupportedLicense() {
+    cy.get('div[class = "form-group alert alert-danger in"]').contains('The selected license is not supported at the moment. Please follow the procedure described under section "None of these licenses suits your needs".');
+  }
+};
+
+describe('Create a new submission', () => {
+  beforeEach(() => {
+    cy.visit('/');
+    // Login as admin
+    loginProcess.login(TEST_ADMIN_USER, TEST_ADMIN_PASSWORD);
+
+    // Create a new submission
+    cy.visit('/submit?collection=' + TEST_SUBMIT_COLLECTION_UUID + '&entityType=none');
+  });
+
+  it('should select the license from the license selection dropdown and change status', {
+    retries: {
+      runMode: 6,
+      openMode: 6,
+    },
+    defaultCommandTimeout: 10000
+  }, () => {
+    createItemProcess.checkLicenseResourceStep();
+    // check default value in the license dropdown selection
+    createItemProcess.checkLicenseSelectionValue('Select a License ...');
+    // check step status - it should be as warning
+    createItemProcess.checkResourceLicenseStatus('Warnings');
+    // select `Public Domain Mark (PD)` from the selection
+    createItemProcess.selectValueFromLicenseSelection('Public Domain Mark (PD)');
+    // selected value should be seen as selected value in the selection
+    createItemProcess.checkLicenseSelectionValue('Public Domain Mark (PD)');
+    // check step status - it should be valid
+    createItemProcess.checkResourceLicenseStatus('Valid');
+  });
+});
